+++ src/main/java/hu/kazocsaba/math/matrix/CoredVector3Impl.java
@@ -43,7 +43,12 @@ class CoredVector3Impl extends CoredVectorImpl implements Vector3 {
 
 	@Override
 	public Vector4 toHomogeneous() {
-		return Vector3Op.toHomogeneous(this);
+		return (Vector4)Vector3Op.toHomogeneous(this);
+	}
+
+	@Override
+	public Vector2 fromHomogeneous() {
+		return (Vector2)Vector3Op.fromHomogeneous(this);
 	}
 
 	@Override