+++ src/main/java/hu/kazocsaba/math/matrix/CoredVectorImpl.java
@@ -65,4 +65,9 @@ class CoredVectorImpl extends CoredMatrixImpl implements Vector {
 	public Vector toHomogeneous() {
 		return VectorOp.toHomogeneous(this);
 	}
+
+	@Override
+	public Vector fromHomogeneous() {
+		return VectorOp.fromHomogeneous(this);
+	}
 }