+++ src/main/java/hu/kazocsaba/math/matrix/backbone/Vector3Op.java
@@ -3,7 +3,6 @@ package hu.kazocsaba.math.matrix.backbone;
 import hu.kazocsaba.math.matrix.Matrix;
 import hu.kazocsaba.math.matrix.MatrixFactory;
 import hu.kazocsaba.math.matrix.Vector3;
-import hu.kazocsaba.math.matrix.Vector4;
 
 /**
  * Implementations of the Vector3 functions for the benefit of Vector subclasses. These methods rely on the
@@ -22,12 +21,5 @@ public final class Vector3Op extends VectorOp {
 				caller.getY() * v.getZ() - caller.getZ() * v.getY(),
 				caller.getZ() * v.getX() - caller.getX() * v.getZ(),
 				caller.getX() * v.getY() - caller.getY() * v.getX());
-	}
-
-	/**
-	 * @see Vector3#toHomogeneous
-	 */
-	public static Vector4 toHomogeneous(Vector3 caller) {
-		return MatrixFactory.createVector(caller.getX(), caller.getY(), caller.getZ(), 1);
 	}
 }